+++ kamehouse-groot/src/main/php/kame-house-groot/api/v1/kamehouse/admin/kamehouse-shell/kamehouse-shell.php
@@ -104,10 +104,10 @@ class KameHouseShell {
     global $kameHouse;
     if ($kameHouse->core->isLinuxHost()) {
       $username = $kameHouse->core->getKameHouseUser();
-      return rtrim(shell_exec("sudo /home/" . $username . "/programs/kamehouse-shell/bin/common/sudoers/www-data/su.sh kamehouse/get-kamehouse-secret.sh -s " . $secretKey));
+      return rtrim(shell_exec("sudo /home/" . $username . "/programs/kamehouse-shell/bin/common/sudoers/www-data/su.sh kamehouse/secrets/get-kamehouse-secret.sh -s " . $secretKey));
     } else {
       $shellScriptsBasePath = $this->getShellScriptsBasePath();
-      $shellCommand = "%USERPROFILE%/programs/kamehouse-shell/bin/win/bat/git-bash-silent.bat -c \"~" . $shellScriptsBasePath . "kamehouse/get-kamehouse-secret.sh -s " . $secretKey;
+      $shellCommand = "%USERPROFILE%/programs/kamehouse-shell/bin/win/bat/git-bash-silent.bat -c \"~" . $shellScriptsBasePath . "kamehouse/secrets/get-kamehouse-secret.sh -s " . $secretKey;
       return rtrim(shell_exec($shellCommand));
     }
   }